// src/types/conference.ts
export type SponsorshipLevel = "PLATINUM" | "GOLD" | "SILVER" | "BRONZE" | "PARTNER";

export interface Sponsor {
  id: number;
  name: string;
  description?: string;
  logoUrl?: string;
  websiteUrl?: string;
  sponsorshipLevel: SponsorshipLevel;
  displayOrder: number;
  status: boolean;
}